+++ main.py
@@ -7,7 +7,7 @@ from utilities.consts import *
 from gamemodes.WorldRunner import WorldRunner
 from gamemodes.WesternMaker2 import WesternMaker
 from gamemodes.TreasureProtect import TreasureProtect
-from gamemodes.MainMenu import MainMenu
+from utilities.MenuSystem import MainMenu, GameModeSelect
 from utilities.utilities import GlobalSettings
 
 pygame.display.set_icon(pygame.image.load(os.path.join(ASSETS_DIRECTORY, WINDOW_ICON)))
@@ -19,7 +19,8 @@ screen.set_alpha(None)
 global_config = GlobalSettings()
 
 games = {"worldrunner": WorldRunner, "westernmaker": WesternMaker,
-         "mainmenu": MainMenu, "treasureprotect": TreasureProtect}
+         "mainmenu": MainMenu, "treasureprotect": TreasureProtect,
+         "gameselect": GameModeSelect}
 global_config.next_game = "mainmenu"
 
 while global_config.game_running:
